hub-workday-connector/src/main/java/com/vmware/ws1connectors/workday/web/resources/WorkdayResource.java
/*
 * Project Workday Connector
 * (c) 2019-2020 VMware, Inc. All rights reserved.
 * SPDX-License-Identifier: BSD-2-Clause
 */
package com.vmware.ws1connectors.workday.web.resources;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import org.springframework.util.CollectionUtils;

import java.util.List;

@NoArgsConstructor
@AllArgsConstructor
@Getter
@Setter
@JsonIgnoreProperties(ignoreUnknown = true)
public class WorkdayResource<T> {
    private int total;
    private List<T> data;

    public boolean hasData() {
        return this.total > 0 && !CollectionUtils.isEmpty(this.data);
    }
}
